+++ src/main/java/uts/asd/hsms/model/dao/UserDao.java
@@ -99,24 +99,7 @@ public class UserDao {
         newRecord.put("userRole", userRole);
         collection.insert(newRecord);
     }
-    
-    public Tutorial[] getTutorials() {
-        DBCursor cursor = collection.find();
-        System.out.println("COUNT: " + cursor.count());
-        Tutorial[] tutorials = new Tutorial[cursor.count()];
-        int count = 0;
-        while (cursor.hasNext()) {
-            DBObject result = cursor.next();
-            ObjectId tutorialId = (ObjectId)result.get("_id");
-            String department = (String)result.get("firstName");
-            int grade = (int)result.get("lastName");
-            ObjectId userId = (ObjectId)result.get("department");
-            int tutSize = (int)result.get("email");
-            tutorials[count] = new Tutorial(tutorialId, department, grade, userId, tutSize);
-            count ++;
-        }
-        return tutorials;
-    }
+     
 //    
 //    public User getUser(String userId, String password) throws SQLException {
 //
