+++ src/main/webapp/css/main.css
@@ -129,3 +129,8 @@ div.main {
 .fa{
      color:#4183D7;
 }
+
+/*vincent adding things*/
+.table#tutTable tr:hover {
+    background-color: #FBE5C3;
+}
